Ignored unkown playability
Unkwon will be treated as good. In the future, it may be worth notifying
staff about these songs to give them a chance to mark them as bad.
I understand that there was a very negative reaction to this. Please
keep in mind, most if not all features I work on are documented publicly
for you to comment on. If you see anything you do not like, tell me.